+++ src/features/library/components/LibraryLayout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LibraryLayout';
