+++ src/features/library/components/QuestionSetList/index.ts
@@ -1,0 +1,2 @@
+export { default } from './QuestionSetList';
+export type { QuestionSetListProps } from './QuestionSetList';
